requirements.txt
flask>=2.3.0
requests>=2.31.0
firecrawl-py>=0.0.1
pandas>=2.0.0
python-dotenv>=1.0.0